Add new general scaling methods for specific generals
Up to this point, we had been scaling generals specifically for their
zone, internally in the generals class.  This allows us to specify the
zone for a single general, using only the general id to prevent us
needing extra general information from the client side.

Additionally, we have a method to randomize ally soldiers for a particular
general, too, so that the ally soldiers can be different from the enemy
soldiers.