Don't keep trying to fill a battle with non-existent generals
With the new general shuffle, we ran into an issue where we were trying
to insert a general into a battle, but we didn't have any remaining
generals, so the randomizer would hang.  Now, if we fail to get a general
placed in 20 tries, we just give up trying to place a general.