Add new battle randomizer logic using only ids
In the older version, the battle randomizer took a Generals object, and
pulled ids from it and then did random logic.  There is really no need
for the randomizer to need to know about the Generals object, because it
really only needs ids of generals it needs to randomize.

Beyond that, this method makes it much easier to do special battle
configuration (such as placing generals outside of the standard logic,
like in the upcoming warlord mode), as the client can modify the group
of generals to randomize at will.