Move the Gui Yang battle to the castle entrance
Weird stuff happens because the Gui Yang castle can both contain Zhao Fan
and Zhuge Liang.  With the open ZGL, this can make it impossible to clear
Gui Yang because the ZGL trigger takes precedence.  This also means that
we can't ever move physically past Gui Yang

To get around this, we just move the Gui Yang battle to its entrance and
set all of the Zhao Fan triggers to be already complete.