Make Zhang Bao and Guan Xing immediately available in the new shuffle
In the new shuffling algorithm, we will be randomizing the character here,
but before that matters, we have to make them available.  At the moment,
that makes the current shuffle boring, because we basically have
immediate access to 5 tigers (6 including ZGL), but once we get these
characters randomized, it will be much more fun.