Do a naive replacement of the shuffled NPCs
All these do is replace the NPCs in the NPC data with the shuffled
generals.  This is naive, as many of the generals will need more things
changed (specifically Zhao Yun, as the other generals won't accept the
Zhou Letr).  We moved the Ma Chao logic to the top so as to make sure that
Ma Chao/Ma Dai didn't end up in 2 places.