Improve Zhao Yun randomization
This has 2 parts - the first is that we were using the hardcoded Zhao Yun
scaling, which means that a) Zhao Yun was effectively removed from the game
unless he was randomized into a previous slot, and b) the general at the
Zhao Yun spot would also be a random encounter.

The second part is that in order to get the Zhao Yun character, we have to
give him the Zhou Letr, but the Zhou Letr looks specifically for the Zhao
Yun general Id, so we need to modify that for it to look for the new
general Id.