Fix a bug when Guan Suo gets shuffled
The improved initialization code makes no functional changes from the
vanilla code, it just does it better.  However, one of the special cases
in the init code is setting Guan Suo to C0.  This isn't a big deal unless
he's randomized into a talk-and-join spot, at which point the C0 makes him
not appear.  To fix this with the least amount of modification, we just
set the 2nd starting general to C0 twice, preventing Guan Suo from getting
any special init code.