Reconfigure the Qing Zhou cave for the new general shuffle.
Here, we want to do a few things.  First, we want to make the back of the
cave bridge-locked again.  Second, we want to remove the Yellow Scarves.
Third, we want to close the exit so there's only one way out.  Fourth, we
want to make sure the cave stays open even after the Yellow Scarves are
killed.  Finally, we want to put a new general in the back of the cave in
the middle where the Yellow Scarves were.

We don't set the general here, but we configure it so that putting the
general here only requires one byte.