Configure generals for the Qing Zhou cave.
Here, we pull another general out of the zone pool and place it in the
Qing Zhou cave.  Beyond that, we want to randomize the general that
triggers the bridge.  However, we do have to do the latter in a special
way so that the required general can't be the bridge-locked one.

These type of deadlocks will occur more commonly as we add certain other
gatekeeper things, but for now, this is the only consideration we need to
make.