Make zone 0 generals unfirable
The vanilla way to tell if a general could be fired was to check a list
of generals from 0E:9577-9587 and see if the general was at that list.
Now, we want to be able fire anyone who isn't set to zone 0, but not fire
anyone who is set to zone 0, since there's no way to get them back.

The added benefit of this is that it frees up a few bytes for us to work
with.