Fix the Chen Deng recruitment issue with long names
It turns out that in certain situations, exceptionally long names will
cause the text box to break and just stop the whole text process.  For
Chen Deng, that was preventing us from even accepting his offer to join.
We fix that by swizzling a word to allow us to specify spacing and get
a newline for the general name.