Keep the Han Zhong Trigger informant available.
The vanilla code eliminates the NPC that tells you the general that
triggers the bridge after the Yellow Scarves are vanquished.  On the other
hand, in the randomizer, we always need him available, so we remove the
entrance trigger altogether, along with it's referenced code.

This is structured in a new way so as to have an infrastructure to remove
entrance triggers in future locations as we find issues.