Make the Intro Letr useful
This is a useless item even in vanilla as you don't really gain anything
of value from it.  However, in order to both make it useful and prevent
the Chin battle from locking the gemsword out, we instead move the
gemsword location to Sima Hui's hut, and make it intro letr locked.

In vanilla, the intro letr is taken by Shui Jing, and not by the guard.
The guard requires you to talk to a guy in Chen Cang fortress before she
will let you in.  Instead, we have made it so that Shui Jing says the same
think each time, while the door guard checks for the intro letr and moves
if you have it (or have previously turned it in).

To make this complete, we have to also remove the flag that is set by the
guy in Chen Cang to automatically let us into the hut.