Let Wang Gui travel to Yang Zhou
Because Wang Gui could be locked out after the new Luo Yang is up, which
would also lock out Zhou Chao, we need a means of preventing that.  We
don't actually move Wang Gui to Yang Zhou, as he's still available in
Luo Yang, we just allow him to travel to Yang Zhou, too, so he can be
recruited in either spot.