Make ummet requirements push us out in the direction we came.
We have 5 locations that we need to have special logic for when leaving
1) If we haven't defeated the Si Shui battle, we can't go past that gate
   by simply leaving the courtyard
2) If we haven't defeated Luo Yang, we can't go past the castle by simply
   leaving the courtyard
3) If we haven't defeated Guang Zong, we can't go past the castle by
   simply leaving the outskirts
4) If we haven't uncaged Liu Zheng, we can't go past Luo
5) If we haven't triggered the lady at the bridge to move, we can't go past
   Bo Hai

All of these are true in vanilla, but in vanilla, we can only approach
these locations from one direction until the requirements to pass are met.
However, with the cave shuffle and open map, we end up approaching these
locations from different sides, but we are always getting pushed out to
the vanilla entrance location.  This creates side effects like being able
to travel from Ji Zhou to Nan Yang at the start of the game, but not being
able to go back because we keep getting pushed south out of Bo Hai.

To rectify this, we are keeping track of the direction we are facing when
entering a location (from the world map, we don't want to keep track from
a shop), then when we leave one of these locations, we adjust either the
X or Y coordinate depending upon the direction we went.

This has proved more complicated than anticipated.  There is no space to
add a command to store the location we're facing when our direction is
updated on entering a town.  This is further complicated by the fact that
we are executing this code from $06XX rather than from bank E.  We can't
create subroutines in bank E that can be called, as while this code is
executed, the active bank is bank 7, which had no space available.

To get around the bank 7 lack of space, we remove the final screen from
the game intro, and zero the bytes and code that utilize it.  This frees
up plenty of space, and allows us to add the storage code to bank 7.

Finally, the old code to disable the Si Shui lock overwrote a byte in the
code we're overwriting here, so it actually breaks things if we call that
method after the one written here.  Thus, we've moved the Si Shui unlocking
code to come before we add all of this code.