Replace the Xu Zhe Bridge Trigger general
The only texts changed here are the lady's lines.  There are other NPCs
that discuss him, but because his name is so short, many of those other
ones will exceed the line length limit, which in turn would require a
rewrite of many texts.  One of these days, I'm going to do a text
overhaul, but until then, we'll just let the other NPCs be wrong.