Update general names to be more correct.
The comments in the code explain the complex sections where we have to
swap memory locations between 2 generals, one with a shrinking character
count and one with a growing character count.

Note this is ugly and could be massively improved.  There will be
refactorings once version 1.0.0 is released, but I'm trying to get that
out the door first.