Change the NPC data for the talk-and-fight generals and make them unique
The logic here is that we want to make the talk-and-join generals actually
be the generals you fight in battle.  However, if that general isn't
unique to the talk-and-fight battle, then the general could just disappear
when recruited elsewhere.  We want him to only be available if we defeat
him at the location.

Ma Yuanyi had an odd characteristic where we had to do some special logic
so we could always fight him.  Yuan Shang has some special considerations
to handle too, which are described in the comments.  Pirate and Lu Meng
needed no special logic.

We moved the UpdateGenerals call to after the talk-and-join replacements
because we need it to happen after the Yuan Shang logic is applied, since
it changes his flags.